+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ## v0.1.3
 
 - Add Python3.5 support.
+- Make it possible to create new rsdict with rsdict.
 
 ## v0.1.2